Fixed pickle issue
Didn't realize that it's difficult to use pickle across modules and in different folders. DataAnalysis.py is now in the main directory and should be able to load the evolutionary progressions. The DataAnalysis.py is still very messy, I used this file to make the plots of the paper though I realize they are quite incomprehensible. I added a minimal fitness plotter that will load and plot a single evolutionary run. All you have to do is to run the DataAnalysis file with the proper references to the 'path' and 'filename'